Create sim_afu source copy inside ASE target directory
The script was writing a temporary copy of the source tree in a common
directory in the source tree.  Now that there is an explicit, private
ASE build directory specified the sim_afu copy can go there.

This still doesn't support an application that specifies the sources
explicitly, but it is a step.