Fix unreachable code issue with debug_print statement
The debug_print statement was conditioned on FPGA_DMA_DEBUG, but
this was statically set to 0 in the Makefile.  This caused all
of the debug_print statements to be unreachable code with triggers
a warning similar to the one shown below.

Solution is the remove the define of FPGA_DMA_DEBUG, which causes
the `debug_print` macro to be defined as the empty string.

Statement remains in the code and can manually be enabled for
debugging if necessary, but is not built by default.

---------------------------------------------------------------------------
1 (Local) /build2/gsouther/projs/intel-fpga-afu/dma_afu/sw/fpga_dma.c:104 UNREACH.GEN (3:Warning) Analyze
Code is unreachable
  * fpga_dma.c:104: Infeasible condition: '0' cannot be true.
    * fpga_dma.c:104: Result of expression '0 != 0' is '== 0'
  * fpga_dma.c:104: The code is unreachable.
Current status 'Analyze'