Replace memcpy with memcpy_s to comply with klocwork checks
This replaces memcpy with memcpy_s.  It is not really ideal
because memcpy_s should really use the size of the output
buffer rather than just repeating the transfer size.  However,
it was not obvious in most cases what the size of the output
buffer was so I just repeated the transfer size.

If there are ways to know the output buffer size this should
be changed in those cases.