Change AFU interface type's key from "name" to "class". (#139)
Using "name" was a source of confusion every time I explained the PIM's AFU
JSON.  They would see "afu-top-interface:name" and think that's the module
name, which was a reasonable assumption.  It isn't.  Instead, it is the class
of the interface, defining the ports and not the module name.

For compatibility, afu_platform_config accepts the legacy key "name" as a synonym.