Remove all instances of add-extra-timing-reg-stages from AFU PIM JSON (#140)
This was used in DCP 1.0 as a hack because of the need to add extra buffering
to CCI-P Tx/Rx structs at the PR boundary in order to meet timing.  It is
no longer necessary and will not be needed on future platforms.

The FIU team has agreed that all future platforms will honor the single register
stage CCI-P requirement inside ccip_std_afu().  If extra register stages are
required at the PR boundary on a given platform we will implement them in the
green_bs and by adding extra buffer slots in the FIU to leave room for messages
arriving in the FIU after almost full is received by the AFU.